Move "Ballot Not Found" button into BallotAudit component
So that we can use Formik's resetForm action
