+++ client/src/components/AuditBoard/BallotAudit.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Formik, FormikProps, Field } from 'formik'
 import styled from 'styled-components'
-import { H3, Button } from '@blueprintjs/core'
+import { Button, Colors, H3, H4 } from '@blueprintjs/core'
 import {
   BallotRow,
   ContestCard,
@@ -10,11 +10,41 @@ import {
   LeftCheckboxes,
   RightCheckboxes,
   SubTitle,
+  FlushDivider,
 } from './Atoms'
 import FormButton from '../Atoms/Form/FormButton'
 import { IBallotInterpretation, Interpretation, IContest } from '../../types'
+import { IBallot } from '../JurisdictionAdmin/useBallots'
 import FormField from '../Atoms/Form/FormField'
 import BlockCheckbox from './BlockCheckbox'
+import constructEmptyInterpretation from '../../utils/interpretations'
+
+const BallotMainRow = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  justify-content: space-between;
+  gap: 10px;
+`
+
+const BallotRowValue = styled(H4)`
+  margin-bottom: 0;
+  color: ${Colors.BLACK};
+`
+
+const NotFoundButton = styled(Button)`
+  border-radius: 5px;
+  width: 13.5em;
+  font-weight: 600;
+  &.bp3-button.bp3-large {
+    height: 2em;
+    min-height: auto;
+    font-size: 14px;
+  }
+  @media only screen and (max-width: 767px) {
+    width: auto;
+  }
+`
 
 const NoteField = styled(Field)`
   textarea {
@@ -38,84 +68,129 @@ const SubmitButton = styled(FormButton)`
 `
 
 interface IProps {
+  ballot: IBallot
   contests: IContest[]
   interpretations: IBallotInterpretation[]
-  interpretationsInitializedAt: Date
   setInterpretations: (interpretations: IBallotInterpretation[]) => void
-  previousBallot: () => void
   confirmSelections: (interpretations: IBallotInterpretation[]) => void
+  submitBallotNotFound: () => void
+  previousBallot: () => void
 }
 
 const BallotAudit: React.FC<IProps> = ({
+  ballot,
   contests,
   interpretations,
-  interpretationsInitializedAt,
   setInterpretations,
   confirmSelections,
+  submitBallotNotFound,
   previousBallot,
 }: IProps) => {
   return (
-    <BallotRow>
-      <div className="ballot-main">
-        <SubTitle>Ballot Contests</SubTitle>
-        <Formik
-          initialValues={{
-            interpretations,
-            // Formik only reinitializes if `enableReinitialize` is true and `initialValues`
-            // changes as determined by a deep equality check (reference changes don't trigger
-            // reinitialization). Changing `interpretationsInitializedAt` while leaving
-            // `interpretations` intact accordingly allows consumers to reset the form
-            interpretationsInitializedAt,
-          }}
-          enableReinitialize
-          onSubmit={values => {
-            setInterpretations(values.interpretations)
-            confirmSelections(values.interpretations)
-          }}
-        >
-          {({
-            handleSubmit,
-            values,
-            setFieldValue,
-          }: FormikProps<{ interpretations: IBallotInterpretation[] }>) => {
-            return (
-              <form>
-                {contests.map((contest, i) => (
-                  <BallotAuditContest
-                    key={contest.id}
-                    contest={contest}
-                    interpretation={values.interpretations[i]}
-                    setInterpretation={newInterpretation =>
-                      setFieldValue(`interpretations[${i}]`, newInterpretation)
-                    }
-                  />
-                ))}
-                <ProgressActions>
-                  <SubmitButton
-                    type="submit"
-                    onClick={handleSubmit}
-                    intent="success"
-                    large
-                    disabled={
-                      !(
-                        values.interpretations.filter(
-                          ({ interpretation }) => interpretation != null
-                        ).length > 0
-                      )
-                    }
-                  >
-                    Submit Selections
-                  </SubmitButton>
-                  <Button onClick={previousBallot} intent="none">
-                    Back
-                  </Button>
-                </ProgressActions>
-              </form>
-            )
-          }}
-        </Formik>
-      </div>
-    </BallotRow>
+    <Formik
+      initialValues={{ interpretations }}
+      enableReinitialize
+      onSubmit={values => {
+        setInterpretations(values.interpretations)
+        confirmSelections(values.interpretations)
+      }}
+    >
+      {({
+        handleSubmit,
+        values,
+        setFieldValue,
+        resetForm,
+      }: FormikProps<{ interpretations: IBallotInterpretation[] }>) => {
+        return (
+          <>
+            <BallotMainRow>
+              {ballot.batch.container && (
+                <div>
+                  <SubTitle>Container</SubTitle>
+                  <BallotRowValue>{ballot.batch.container}</BallotRowValue>
+                </div>
+              )}
+              {ballot.batch.tabulator && (
+                <div>
+                  <SubTitle>Tabulator</SubTitle>
+                  <BallotRowValue>{ballot.batch.tabulator}</BallotRowValue>
+                </div>
+              )}
+              <div>
+                <SubTitle>Batch</SubTitle>
+                <BallotRowValue>{ballot.batch.name}</BallotRowValue>
+              </div>
+              <div>
+                <SubTitle>Ballot Number</SubTitle>
+                <BallotRowValue>{ballot.position}</BallotRowValue>
+              </div>
+              {ballot.imprintedId !== undefined && (
+                <div>
+                  <SubTitle>Imprinted ID</SubTitle>
+                  <BallotRowValue>{ballot.imprintedId}</BallotRowValue>
+                </div>
+              )}
+              <div>
+                <NotFoundButton
+                  onClick={() => {
+                    setInterpretations(
+                      contests.map(constructEmptyInterpretation)
+                    )
+                    resetForm()
+                    submitBallotNotFound()
+                  }}
+                  intent="danger"
+                  large
+                >
+                  Ballot Not Found
+                </NotFoundButton>
+              </div>
+            </BallotMainRow>
+            <FlushDivider />
+            <BallotRow>
+              <div className="ballot-main">
+                <SubTitle>Ballot Contests</SubTitle>
+                <form>
+                  {contests.map((contest, i) => (
+                    <BallotAuditContest
+                      key={contest.id}
+                      contest={contest}
+                      interpretation={values.interpretations[i]}
+                      setInterpretation={newInterpretation =>
+                        setFieldValue(
+                          `interpretations[${i}]`,
+                          newInterpretation
+                        )
+                      }
+                    />
+                  ))}
+                  <ProgressActions>
+                    <SubmitButton
+                      type="submit"
+                      onClick={handleSubmit}
+                      intent="success"
+                      large
+                      disabled={
+                        !(
+                          values.interpretations.filter(
+                            ({ interpretation }) => interpretation != null
+                          ).length > 0
+                        )
+                      }
+                    >
+                      Submit Selections
+                    </SubmitButton>
+                    <Button onClick={previousBallot} intent="none">
+                      Back
+                    </Button>
+                  </ProgressActions>
+                </form>
+              </div>
+            </BallotRow>
+          </>
+        )
+      }}
+    </Formik>
   )
 }
