+++ client/src/utils/interpretations.ts
@@ -1,0 +1,12 @@
+import { IBallotInterpretation, IContest } from '../types'
+
+export default function constructEmptyInterpretation(
+  contest: IContest
+): IBallotInterpretation {
+  return {
+    contestId: contest.id,
+    interpretation: null,
+    choiceIds: [],
+    comment: null,
+  }
+}
